refactor: rendershapes function
